# Core: Support TS3.8+ again