--- a/code/renderers/react/src/preview.tsx
+++ b/code/renderers/react/src/preview.tsx
@@ -66,7 +66,7 @@ interface ReactMeta<
   ): ReactStory;
 
   story<
-    const TInput extends Simplify<
+    TInput extends Simplify<
       StoryAnnotations<
         ReactRenderer,
         // TODO: infer mocks from story itself as well