Closes #

<!-- If your PR is related to an issue, provide the number(s) above; if it resolves multiple issues, be sure to break them up (e.g. "closes #1000, closes #1001"). -->

<!--

Thank you for contributing to Storybook! Please submit all PRs to the `next` branch unless they are specific to the current release. Storybook maintainers cherry-pick bug and documentation fixes into the `main` branch as part of the release process, so you shouldn't need to worry about this. For additional guidance: https://storybook.js.org/docs/contribute

-->

## What I did

<!-- Briefly describe what your PR does -->

## Checklist for Contributors

### Testing

<!-- Please check (put an "x" inside the "[ ]") the applicable items below to communicate how to test your changes -->

#### The changes in this PR are covered in the following automated tests:

- [ ] stories
- [ ] unit tests
- [ ] integration tests
- [ ] end-to-end tests

#### Manual testing

_This section is mandatory for all contributions. If you believe no manual test is necessary, please state so explicitly. Thanks!_

<!-- Please include the steps to test your changes here. For example:

1. Run a sandbox for template, e.g. `yarn task --task sandbox --start-from auto --template react-vite/default-ts`
2. Open Storybook in your browser
3. Access X story

-->

### Documentation

<!-- Please check (put an "x" inside the "[ ]") the applicable items below to indicate which documentation has been updated. -->

- [ ] Add or update documentation reflecting your changes
- [ ] If you are deprecating/removing a feature, make sure to update
      [MIGRATION.MD](https://github.com/storybookjs/storybook/blob/next/MIGRATION.md)

## Checklist for Maintainers

- [ ] When this PR is ready for testing, make sure to add `ci:normal`, `ci:merged` or `ci:daily` GH label to it to run a specific set of sandboxes. The particular set of sandboxes can be found in `code/lib/cli-storybook/src/sandbox-templates.ts`
- [ ] Make sure this PR contains **one** of the labels below:
   <details>
     <summary>Available labels</summary>

  - `bug`: Internal changes that fixes incorrect behavior.
  - `maintenance`: User-facing maintenance tasks.
  - `dependencies`: Upgrading (sometimes downgrading) dependencies.
  - `build`: Internal-facing build tooling & test updates. Will not show up in release changelog.
  - `cleanup`: Minor cleanup style change. Will not show up in release changelog.
  - `documentation`: Documentation **only** changes. Will not show up in release changelog.
  - `feature request`: Introducing a new feature.
  - `BREAKING CHANGE`: Changes that break compatibility in some way with current major version.
  - `other`: Changes that don't fit in the above categories.

   </details>

### 🦋 Canary release

<!-- CANARY_RELEASE_SECTION -->

This PR does not have a canary release associated. You can request a canary release of this pull request by mentioning the `@storybookjs/core` team here.

_core team members can create a canary release [here](https://github.com/storybookjs/storybook/actions/workflows/canary-release-pr.yml) or locally with `gh workflow run --repo storybookjs/storybook canary-release-pr.yml --field pr=<PR_NUMBER>`_

<!-- CANARY_RELEASE_SECTION -->

<!-- BENCHMARK_SECTION -->

| name | before | after | diff | z | % |
| ---- | ------ | ----- | ---- | - | - |
| createSize |  0 B | 0 B | 0 B | - | - |
| generateSize |  79.7 MB | 79.7 MB | 0 B | -0.48 | 0% |
| initSize |  79.7 MB | 79.7 MB | 0 B | -0.48 | 0% |
| diffSize |  97 B | 97 B | 0 B | - | 0% |
| buildSize |  7.31 MB | 7.31 MB | 0 B | 0.65 | 0% |
| buildSbAddonsSize |  1.9 MB | 1.9 MB | 0 B | 0.43 | 0% |
| buildSbCommonSize |  195 kB | 195 kB | 0 B | - | 0% |
| buildSbManagerSize |  1.89 MB | 1.89 MB | 0 B | 0.31 | 0% |
| buildSbPreviewSize |  0 B | 0 B | 0 B | - | - |
| buildStaticSize |  0 B | 0 B | 0 B | - | - |
| buildPrebuildSize |  3.98 MB | 3.98 MB | 0 B | **1.28** | 0% |
| buildPreviewSize |  3.33 MB | 3.33 MB | 0 B | 0.24 | 0% |
| testBuildSize |  0 B | 0 B | 0 B | - | - |
| testBuildSbAddonsSize |  0 B | 0 B | 0 B | - | - |
| testBuildSbCommonSize |  0 B | 0 B | 0 B | - | - |
| testBuildSbManagerSize |  0 B | 0 B | 0 B | - | - |
| testBuildSbPreviewSize |  0 B | 0 B | 0 B | - | - |
| testBuildStaticSize |  0 B | 0 B | 0 B | - | - |
| testBuildPrebuildSize |  0 B | 0 B | 0 B | - | - |
| testBuildPreviewSize |  0 B | 0 B | 0 B | - | - |



| name | before | after | diff | z | % |
| ---- | ------ | ----- | ---- | - | - |
| createTime |  7.6s | 7s | -597ms | -0.97 | -8.4% |
| generateTime |  22.2s | 17.8s | -4s -456ms | -1.09 | -25% |
| initTime |  5.3s | 4s | -1s -333ms | **-1.39** | 🔰-32.9% |
| buildTime |  11.6s | 8s | -3s -543ms | **-1.42** | 🔰-43.9% |
| testBuildTime |  0ms | 0ms | 0ms | - | - |
| devPreviewResponsive |  5.1s | 5.6s | 433ms | 0.27 | 7.7% |
| devManagerResponsive |  4.9s | 5.3s | 350ms | 0.24 | 6.6% |
| devManagerHeaderVisible |  655ms | 973ms | 318ms | **1.27** | 🔺32.7% |
| devManagerIndexVisible |  697ms | 1s | 313ms | 1.18 | 31% |
| devStoryVisibleUncached |  1.2s | 2.2s | 1s | 1.01 | 45.5% |
| devStoryVisible |  724ms | 1s | 287ms | 0.95 | 28.4% |
| devAutodocsVisible |  707ms | 862ms | 155ms | 0.69 | 18% |
| devMDXVisible |  670ms | 882ms | 212ms | 1.04 | 24% |
| buildManagerHeaderVisible |  645ms | 1.1s | 535ms | **2.96** | 🔺45.3% |
| buildManagerIndexVisible |  654ms | 1.1s | 545ms | **2.4** | 🔺45.5% |
| buildStoryVisible |  605ms | 983ms | 378ms | **1.75** | 🔺38.5% |
| buildAutodocsVisible |  586ms | 622ms | 36ms | -0.22 | 5.8% |
| buildMDXVisible |  505ms | 719ms | 214ms | 1.06 | 29.8% |

<!-- BENCHMARK_SECTION -->


<!-- greptile_comment -->

## Greptile Summary

This PR restores TypeScript 3.8+ compatibility across the Storybook codebase by renaming and restructuring how `definePreview` is exported from framework packages.

- Renamed `definePreview` to `__definePreview` in `code/core/src/csf/csf-factories.ts` with appropriate deprecation notices
- Updated all framework packages to import `__definePreview` and re-export it as `definePreview` for consistent API
- Enhanced `cleanupTypeImports` in `code/lib/cli-storybook/src/codemod/helpers/csf-factories-utils.ts` to track used identifiers and prevent removal of imports used elsewhere
- Simplified documentation by adding `hideRendererSelector: true` to framework-specific pages
- Optimized CI configuration by reducing parallelism counts for several jobs

Greptile AI



<!-- /greptile_comment -->